split hot_region.go
